+++ src/Plugin/GraphQL/DataProducer/EntityQuery.php
@@ -2,8 +2,12 @@
 
 namespace Drupal\thunder_gqls\Plugin\GraphQL\DataProducer\Entity;
 
-use Drupal\graphql\GraphQL\Execution\FieldContext;
-use Drupal\thunder_gqls\Plugin\GraphQL\DataProducer\ThunderEntityQueryBase;
+use Drupal\Core\Cache\RefinableCacheableDependencyInterface;
+use Drupal\Core\Entity\EntityTypeManager;
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
+use Drupal\graphql\Plugin\GraphQL\DataProducer\DataProducerPluginBase;
+use GraphQL\Error\UserError;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Builds and executes Drupal entity query.
@@ -12,92 +16,97 @@ use Drupal\thunder_gqls\Plugin\GraphQL\DataProducer\ThunderEntityQueryBase;
  *   id = "thunder_entity_query",
  *   name = @Translation("Load entities"),
  *   description = @Translation("Loads entities."),
- *   produces = @ContextDefinition("string",
- *     label = @Translation("Entity IDs"),
- *     multiple = TRUE
+ *   produces = @ContextDefinition("entities",
+ *     label = @Translation("Entities")
  *   ),
  *   consumes = {
  *     "type" = @ContextDefinition("string",
  *       label = @Translation("Entity type")
  *     ),
+ *     "bundles" = @ContextDefinition("any",
+ *       label = @Translation("Entity bundle(s)"),
+ *       multiple = TRUE,
+ *       required = FALSE
+ *     ),
  *     "limit" = @ContextDefinition("integer",
  *       label = @Translation("Limit"),
- *       required = FALSE,
- *       default_value = 10
+ *       required = FALSE
  *     ),
  *     "offset" = @ContextDefinition("integer",
  *       label = @Translation("Offset"),
- *       required = FALSE,
- *       default_value = 0
- *     ),
- *     "owned_only" = @ContextDefinition("boolean",
- *       label = @Translation("Query only owned entities"),
- *       required = FALSE,
- *       default_value = FALSE
+ *       required = FALSE
  *     ),
  *     "conditions" = @ContextDefinition("any",
  *       label = @Translation("Conditions"),
  *       multiple = TRUE,
- *       required = FALSE,
- *       default_value = {}
- *     ),
- *     "allowed_filters" = @ContextDefinition("string",
- *       label = @Translation("Allowed filters"),
- *       multiple = TRUE,
- *       required = FALSE,
- *       default_value = {}
- *     ),
- *     "languages" = @ContextDefinition("string",
- *       label = @Translation("Entity languages"),
- *       multiple = TRUE,
- *       required = FALSE,
- *       default_value = {}
- *     ),
- *     "bundles" = @ContextDefinition("any",
- *       label = @Translation("Entity bundles"),
- *       multiple = TRUE,
- *       required = FALSE,
- *       default_value = {}
- *     ),
- *     "sorts" = @ContextDefinition("any",
- *       label = @Translation("Sorts"),
- *       multiple = TRUE,
- *       default_value = {},
  *       required = FALSE
  *     )
  *   }
  * )
  */
-class EntityQuery extends ThunderEntityQueryBase {
+class EntityQuery extends DataProducerPluginBase implements ContainerFactoryPluginInterface {
 
   /**
-   * The default maximum number of items to be capped to prevent DDOS attacks.
+   * Maximum number of results.
+   *
+   * To prevent denial of service attacks with loading too many items.
    */
-  const MAX_ITEMS = 100;
+  const SIZE_MAX = 50;
+
+  /**
+   * The entity type manager service.
+   *
+   * @var \Drupal\Core\Entity\EntityTypeManager
+   */
+  protected $entityTypeManager;
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
+    return new static(
+      $configuration,
+      $plugin_id,
+      $plugin_definition,
+      $container->get('entity_type.manager')
+    );
+  }
+
+  /**
+   * EntityLoad constructor.
+   *
+   * @param array $configuration
+   *   The plugin configuration array.
+   * @param string $pluginId
+   *   The plugin id.
+   * @param array $pluginDefinition
+   *   The plugin definition array.
+   * @param \Drupal\Core\Entity\EntityTypeManager $entityTypeManager
+   *   The entity type manager service.
+   */
+  public function __construct(
+    array $configuration,
+    string $pluginId,
+    array $pluginDefinition,
+    EntityTypeManager $entityTypeManager
+  ) {
+    parent::__construct($configuration, $pluginId, $pluginDefinition);
+    $this->entityTypeManager = $entityTypeManager;
+  }
 
   /**
    * Resolves the entity query.
    *
    * @param string $type
    *   Entity type.
+   * @param array|null $bundles
+   *   List of bundles to be filtered.
    * @param int $limit
    *   Maximum number of queried entities.
-   * @param int $offset
+   * @param int|null $offset
    *   Offset to start with.
-   * @param bool $ownedOnly
-   *   Query only entities owned by current user.
-   * @param array $conditions
+   * @param array|null $conditions
    *   List of conditions to filter the entities.
-   * @param array $allowedFilters
-   *   List of fields to be used in conditions to restrict access to data.
-   * @param string[] $languages
-   *   Languages for queried entities.
-   * @param string[] $bundles
-   *   List of bundles to be filtered.
-   * @param array $sorts
-   *   List of sorts.
-   * @param \Drupal\graphql\GraphQL\Execution\FieldContext $context
-   *   The caching context related to the current field.
    *
    * @return array
    *   The list of ids that match this query.
@@ -105,50 +114,42 @@ class EntityQuery extends ThunderEntityQueryBase {
    * @throws \GraphQL\Error\UserError
    *   No bundles defined for given entity type.
    */
-  public function resolve(string $type, int $limit, int $offset, bool $ownedOnly, array $conditions, array $allowedFilters, array $languages, array $bundles, array $sorts, FieldContext $context): array {
-    $storage = $this->entityTypeManager->getStorage($type);
-    $entityType = $storage->getEntityType();
+  public function resolve(string $type, ?array $bundles = NULL, int $limit = 10, ?int $offset = NULL, ?array $conditions = NULL): array {
 
-    $query = $this->buildBaseEntityQuery(
-      $type,
-      $ownedOnly,
-      $conditions,
-      $allowedFilters,
-      $languages,
-      $bundles,
-      $context
-    );
+    // Make sure that max limit is not crossed.
+    if ($limit > static::SIZE_MAX) {
+      $limit = static::SIZE_MAX;
+    }
 
     // Make sure offset is zero or positive.
-    $offset = max($offset, 0);
-
-    // Make sure limit is positive and cap the max items to prevent DDOS
-    // attacks.
-    if ($limit <= 0) {
-      $limit = 10;
+    if (!isset($offset) || $offset < 0) {
+      $offset = 0;
     }
-    $limit = min($limit, self::MAX_ITEMS);
 
-    // Apply offset and limit.
-    $query->range($offset, $limit);
+    $entity_type = $this->entityTypeManager->getStorage($type);
+    $query = $entity_type->getQuery()
+      ->range($offset, $limit);
 
-    // Add sorts.
-    foreach ($sorts as $sort) {
-      if (!empty($sort['field'])) {
-        if (!empty($sort['direction']) && strtolower($sort['direction']) == 'desc') {
-          $direction = 'DESC';
-        }
-        else {
-          $direction = 'ASC';
-        }
-        $query->sort($sort['field'], $direction);
+    // Ensure that access checking is performed on the query.
+    $query->accessCheck(TRUE);
+
+    if (isset($bundles)) {
+      $bundle_key = $entity_type->getEntityType()->getKey('bundle');
+      if (!$bundle_key) {
+        throw new UserError('No bundles defined for given entity type.');
       }
+      $query->condition($bundle_key, $bundles, "IN");
+    }
+    if (isset($language)) {
+      $query->condition('langcode', $language);
     }
 
-    $metadata->addCacheTags($entityType->getListCacheTags());
-    $metadata->addCacheContexts($entityType->getListCacheContexts());
+    foreach ($conditions as $condition) {
+      $operation = isset($condition['operator']) ? $condition['operator'] : NULL;
+      $query->condition($condition['field'], $condition['value'], $operation);
+    }
 
-    return new QueryConnection($query);
+    $ids = $query->execute();
 
     return $ids;
   }
